Refactor mocks into the test-config folder
