+++ src/app/(frontend)/admin/page.tsx
@@ -6,9 +6,9 @@ import NavBar from '@/components/Generic/NavBar/NavBar'
 import { UniqueIdentifier } from '@dnd-kit/core'
 import { useEffect, useState } from 'react'
 import { motion } from 'framer-motion'
-import { mockProjects } from '@/mocks/newProjects.mock'
-import { clientMocks as mockClients } from '@/mocks/newClients.mock'
-import { mockSemesters } from '@/mocks/newSemesters.mock'
+import { mockProjects } from '@/test-config/mocks/Project.mock'
+import { clientMocks as mockClients } from '@/test-config/mocks/User.mock'
+import { mockSemesters } from '@/test-config/mocks/Semester.mock'
 import SadTeapot from 'src/assets/sad-teapot.svg'
 import { handleLoginButtonClick, isLoggedIn } from '@/lib/services/user/Handlers'
