+++ src/components/Composite/ClientGroup/ClientGroup.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react'
 import ClientGroup from './ClientGroup'
-import { clientMocks } from '@/mocks/newClients.mock'
+import { clientMocks } from '@/test-config/mocks/User.mock'
 
 const meta: Meta<typeof ClientGroup> = {
   component: ClientGroup,
