+++ src/components/Composite/ClientModal/ClientModal.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from '@storybook/react'
 import { useArgs } from '@storybook/preview-api'
 import ClientModal from './ClientModal'
 import Button from '@/components/Generic/Button/Button'
-import { mockProjects } from '@/mocks/newProjects.mock'
+import { mockProjects } from '@/test-config/mocks/Project.mock'
 
 const meta: Meta<typeof ClientModal> = {
   title: 'Composite/ClientModal',
