+++ src/components/Composite/ClientProfile/ClientProfile.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react'
 import ClientProfile from './ClientProfile'
-import { clientMocks } from '@/mocks/newClients.mock'
+import { clientMocks } from '@/test-config/mocks/User.mock'
 
 const meta: Meta<typeof ClientProfile> = {
   component: ClientProfile,
