--- src/mocks/newClients.mock.tsx
@@ -1,37 +1,0 @@
-import { UserCombinedInfo } from '@/types/Collections'
-
-export const clientMocks: UserCombinedInfo[] = [
-  {
-    id: '1',
-    firstName: 'David',
-    lastName: 'Cumin',
-    email: 'davidcumin@gmail.com',
-    affiliation: 'University of Auckland',
-    introduction: 'Hello nice to meet you! I am David.',
-    role: 'client',
-    createdAt: '2023-10-01T12:00:00Z',
-    updatedAt: '2023-10-01T12:00:00Z',
-  },
-  {
-    id: '2',
-    firstName: 'Dole',
-    lastName: 'Paprika',
-    email: 'dolepaprika@auckland.ac.nz',
-    affiliation: 'University of Auckland',
-    introduction: 'Hello nice to meet you!',
-    role: 'client',
-    createdAt: '2023-10-01T12:00:00Z',
-    updatedAt: '2023-10-01T12:00:00Z',
-  },
-  {
-    id: '3',
-    firstName: 'Dana',
-    lastName: 'Rosemary',
-    email: 'danarosemary@gmail.com',
-    affiliation: 'University of Auckland',
-    introduction: 'Hello nice to meet you! I am Dana.',
-    role: 'client',
-    createdAt: '2023-10-01T12:00:00Z',
-    updatedAt: '2023-10-01T12:00:00Z',
-  },
-]
