+++ src/test-config/mocks/User.mock.ts
@@ -1,5 +1,6 @@
 import { CreateClientAdditionalInfoData, CreateUserData } from '@/types/Collections'
 import { UserRole } from '@/types/User'
+import { UserCombinedInfo } from '@/types/Collections'
 
 export const adminCreateMock: CreateUserData = {
   role: UserRole.Admin,
@@ -27,3 +28,39 @@ export const clientAdditionalInfoCreateMock: CreateClientAdditionalInfoData = {
   introduction: "Hello i'm a very cool client. My project cool.",
   affiliation: 'Very cool company',
 }
+
+export const clientMocks: UserCombinedInfo[] = [
+  {
+    id: '1',
+    firstName: 'David',
+    lastName: 'Cumin',
+    email: 'davidcumin@gmail.com',
+    affiliation: 'University of Auckland',
+    introduction: 'Hello nice to meet you! I am David.',
+    role: 'client',
+    createdAt: '2023-10-01T12:00:00Z',
+    updatedAt: '2023-10-01T12:00:00Z',
+  },
+  {
+    id: '2',
+    firstName: 'Dole',
+    lastName: 'Paprika',
+    email: 'dolepaprika@auckland.ac.nz',
+    affiliation: 'University of Auckland',
+    introduction: 'Hello nice to meet you!',
+    role: 'client',
+    createdAt: '2023-10-01T12:00:00Z',
+    updatedAt: '2023-10-01T12:00:00Z',
+  },
+  {
+    id: '3',
+    firstName: 'Dana',
+    lastName: 'Rosemary',
+    email: 'danarosemary@gmail.com',
+    affiliation: 'University of Auckland',
+    introduction: 'Hello nice to meet you! I am Dana.',
+    role: 'client',
+    createdAt: '2023-10-01T12:00:00Z',
+    updatedAt: '2023-10-01T12:00:00Z',
+  },
+]
